fix cycle dependency
